requirements.txt
# ============================================================
# HybridQA-Net — зависимости
# Python 3.9+
# ============================================================

# ----- Deep Learning ----------------------------------------
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0                  # Vision Transformer, EfficientNet

# ----- HuggingFace ------------------------------------------
transformers>=4.35.0         # RuBERT, LaBSE, T5
tokenizers>=0.15.0
sentencepiece>=0.1.99        # T5 токенизатор

# ----- Computer Vision --------------------------------------
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# ----- NLP / Document parsing -------------------------------
pypdf>=3.0.0                 # PDF парсинг

# ----- API --------------------------------------------------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6      # File uploads
pydantic>=2.4.0

# ----- Auth -------------------------------------------------
python-jose[cryptography]>=3.3.0   # JWT

# ----- Configuration ----------------------------------------
PyYAML>=6.0.1

# ----- Cache (опционально) ----------------------------------
redis>=5.0.0                 # Только если используется Redis backend

# ----- Utilities --------------------------------------------
tqdm>=4.66.0

# ----- Testing ----------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0                # Для тестирования FastAPI
